requirements.txt
requests
beautifulsoup4
pycryptodome
prettytable
ddddocr